Added Day6 of 30days of code
I have now started adding the question now. Adding the questions makes it easier for the viewers as well as my future self to understand the code properly and also adding the question is convenient for the users as now they don't have to keep going back to Hackerrank to take reference of the question that I have solved.